==> Jupyter_notebooky/00 Vybudovanie databázy CRZ.ipynb <==
### Ako používať Jupyter Notebooky? ###
Všetky nástroje a skripty sa dajú spustiť stlačením **Shift+Enter** po vybratí konkrétnej bunky.

### Čo robí tento Jupyter notebook ###
Nasledujúci kód stiahne celú databázu CRZ v XML súboroch, ktoré uloží do priečinku **CRZ_DB**, v ktorom beží Jupyter notebook.

import urllib.request
import zipfile
import os

from datetime import date
from dateutil.rrule import rrule, DAILY

working_dir   = os.getcwd()+'\\CRZ_DB\\'
if not os.path.exists('CRZ_DB'):
    os.makedirs('CRZ_DB')

start_date = date(2011, 1, 1)
end_date   = date(2019, 7, 15)

dates      = []

for dt in rrule(DAILY, dtstart=start_date, until=end_date):
    dates.append(dt.strftime("%Y-%m-%d"))

for date in dates:
    print('Downloading date : '+date)

    # Download
    urllib.request.urlretrieve('http://www.crz.gov.sk//export/'+date+'.zip', os.path.join(working_dir,date+'.zip'))

    # Unzip
    zip_ref = zipfile.ZipFile(date+'.zip', 'r')
    zip_ref.extractall('')
    zip_ref.close()

    # Delete
    os.system('del '+date+'.zip')

Zo stiahnutých XML súborov vybudujeme štrukturovanú tabuľku s názvom **CRZ_DB.csv**. Táto tabuľka ešte neobsahuje dodatky, tie vyriešime nižšie.

import os
import xml.etree.cElementTree as ET
import numpy as np
import pandas as pd

working_dir   = os.getcwd()+'\\CRZ_DB\\'
corrupted_dir = os.getcwd()+'\\Corrupted_XML_files\\'
files    = [f for f in os.listdir(working_dir) if os.path.isfile(os.path.join(working_dir, f))]

table = []

for f in files:
    try:
        file = ET.parse(working_dir+f)
        contracts = file.getroot()
    
        print('Parsing file ... '+f)

        if (len(list(contracts)) > 0):

            for contract in contracts:
                contract_name = contract[4].text
                contract_ID   = contract[1].text

                contract_inner_ID = contract[0].text

                contract_purchaser         = contract[2].text
                contract_purchaser_address = contract[21].text
                contract_purchaser_ICO     = contract[20].text

                contract_supplier         = contract[3].text
                contract_supplier_address = contract[19].text
                contract_supplier_ICO     = contract[13].text

                contract_date_publication = contract[12].text
                contract_date_signed      = contract[24].text
                contract_date_validity    = contract[6].text
                contract_date_efficiency  = contract[5].text
                contract_date_last_change = contract[16].text

                contract_price_final  = contract[8].text
                contract_price_signed = contract[7].text

                contract_resort = contract[11].text

                contract_type  = contract[23].text
                contract_state = contract[14].text

                contract_attachments = []
                attachments = contract[32]

                for attachment in attachments:
                    contract_attachment_ID    = attachment[0].text
                    contract_attachment_name  = attachment[1].text

                    contract_attachment_document_scan = attachment[2].text
                    contract_attachment_size_scan     = int(attachment[3].text)

                    contract_attachment_document_text = attachment[4].text
                    contract_attachment_size_text     = int(attachment[5].text)

                    contract_text = False
                    suffix = ''

                if (contract_attachment_size_text != 0 ):
                        suffix   = '&text=1'
                        contract_text = True

                    contract_attachment_PDF  = 'https://www.crz.gov.sk/index.php?ID=603&doc=' + attachment[0].text + suffix

                    contract_attachment_size = str(max(contract_attachment_size_scan, contract_attachment_size_text))
                    contract_attachment_date = attachment[6].text

                contract_attachments.append([contract_attachment_ID, contract_attachment_name, contract_attachment_PDF, contract_attachment_size, contract_attachment_date, contract_text])

                table.append([contract_name, contract_ID, contract_inner_ID, contract_purchaser_ICO, contract_purchaser, contract_purchaser_address,
                            contract_supplier_ICO, contract_supplier, contract_supplier_address, contract_date_publication, contract_date_signed, contract_date_validity, contract_date_efficiency,
                            contract_date_last_change, contract_price_final, contract_price_signed, contract_resort, contract_type, contract_state, contract_attachments])
    except:
        os.system('move '+working_dir+f+' '+corrupted_dir+f)

header = ['Nazov','ID','Inner-ID','Objednavatel_ICO','Objednavatel','Objednavatel_adresa','Dodavatel_ICO','Dodavatel','Dodavatel_adresa',
             'Datum_zverejnenia','Datum_podpisu','Datum_platnosti','Datum_ucinnosti','Posledna_zmena','Cena_konecna','Cena_podpisana','Rezort','Typ','Stav','Prilohy']

table = np.asarray(table, dtype='object')
# Pandas export better to UTF-8 CSV than raw NumPy
pd.DataFrame(table).to_csv('CRZ_DB.csv', header = header, sep='|')

CRZ neexportuje dodatky v súboroch, ktoré dodáva štandardne. Dodatky sú zobrazované na samostatnej stránke, nasledujúci skript nájde všetky ID kódy dodatkov, ktoré neskôr spáruje k zmluvám.

import requests
import lxml.html as lh
import pandas as pd
import re

find_ID = re.compile(r'\d+')

# Find last starting ID --> 1 000 000 000 is just sufficiently large number
url  = 'https://www.crz.gov.sk/index.php?ID=114430&page=1000000000'

page = requests.get(url)
doc  = lh.fromstring(page.content)

last_ID = int(doc.find_class('pagelist')[0][-1].text_content())

print('Up to this date there are ',last_ID*20,'supplemental agreements ...')
print('Going to crawl CRZ GOV and build DB_supplements.')

supplements_ID = []
page_ID = 1
for page_ID in range(0,last_ID):
    print('\tProcessing page',page_ID,'out of',last_ID)
    url  = 'https://www.crz.gov.sk/index.php?ID=114430&page='+str(page_ID)

    page = requests.get(url)
    doc  = lh.fromstring(page.content)

    tr_elements = doc.xpath('//tr')
    supplements = [supplement for supplement in tr_elements if len(supplement) == 5]
    IDs = [find_ID.findall(supplement[1][0].attrib['href'])[0] for supplement in supplements[1:]]

    supplements_ID = supplements_ID + IDs

    f = open('IDs.txt', 'a')
    for ID in IDs:
        f.write(ID+'\n')
    f.close()

Na základe ID dodatkov vybudujeme štrukturovanú tabuľku **CRZ_DB_supplements.csv**

import requests
import lxml.html as lh
import pandas as pd
import re

find_price_dot         = re.compile(r'\d+\.\d+')
find_price_without_dot = re.compile(r'\d+')

def find_price(string):
    if (len(find_price_dot.findall(string))>0):
        return float(find_price_dot.findall(string)[0])
    else:
        return float(find_price_without_dot.findall(string)[0])

find_ID = re.compile(r'\d+')
header  = ['Nazov','ID_supplement','ID_zmluva','Inner-ID','Objednavatel','Dodavatel','Datum_podpisu','Datum_platnosti','Datum_ucinnosti','Poznamka','Prilohy']

fo  = open('IDs.txt','r')
IDs = fo.readlines()
fo.close()

row_list = []
print('Going to build supplemental agreements DB of',len(IDs))

for i, ID in enumerate(IDs):
    try:
        ID = ID.strip()
        print('Processing ID:',ID,' ',i+1,'out of',len(IDs))
        url = 'https://www.crz.gov.sk/index.php?ID='+ID+'&l=sk'

        page = requests.get(url)
        doc  = lh.fromstring(page.content)
    
        # Metadata about price and dates
        dates_area = doc.find_class('area area1')[0][0][0]
        if (len(dates_area) == 4):
            date_signed     = dates_area[0][1].text_content()
            date_efficiency = dates_area[1][1].text_content()
            date_validity   = dates_area[2][1].text_content()
            price           = find_price(dates_area[3][1].text_content())
        else:
            date_signed     = 'neuvedené'
            date_efficiency = dates_area[0][1].text_content()
            date_validity   = dates_area[1][1].text_content()
            price           = find_price(dates_area[2][1].text_content())

        # Metadata about name, number, supplier and purchaser
        text_area  = doc.find_class('b_right area area3')[0][1][0]
        supplement_number    = text_area[0][1].text_content()
        supplement_purchaser = text_area[1][1].text_content()
        supplement_supplier  = text_area[2][1].text_content()
        supplement_name      = text_area[3][1].text_content()
        if (len(text_area) == 5):
            supplement_note = text_area[4][1].text_content()
        else:
            supplement_note = ''

        # Link to the contract ID in CRZ.gov.sk
        contract_ID  = find_ID.findall(doc.find_class('area5')[0][0].attrib['href'])[0]

        # Supplement attachments
        supplement_attachments = []
        attachments = doc.find_class('area area2')[0][1]
        for attachment in attachments:

            attachment_link = 'https://www.crz.gov.sk'+attachment[1].attrib['href']
            attachment_name = attachment[1].text_content()

            if 'Text' in attachment[0].attrib['alt']:
                attachment_text = True
            else:
                attachment_text = False

            supplement_attachments.append([attachment_link,attachment_name,attachment_text])

        data   = [supplement_name, ID, contract_ID, supplement_number, supplement_purchaser, supplement_supplier, date_signed, date_validity, date_efficiency, supplement_note, supplement_attachments]
        row_list.append(dict((label,data[i]) for i, label in enumerate(header)))

    except:
        pass

    if (i % 50 == 0):
        DB = pd.DataFrame(row_list, columns = header)
        DB.to_csv('CRZ_DB_supplements.csv', header = header, sep = '|')

Posledným krokom je spojenie tabuľky získanej priamo z CRZ z XML súborov z pracne vybudovanou tabuľkou dodatkov a spárovanie dodatkov k jednotlivým zmluvám. Tá je pomenovaná **CRZ_DB_with_supplements.csv**.

import requests
import lxml.html as lh
import pandas as pd
import re

DB             = pd.read_csv('CRZ_DB.csv', sep = '|')
DB_supplements = pd.read_csv('CRZ_DB_supplements.csv', sep = '|')

DB['Dodatky']  = ''

# Structure dictionary (ID_contract,[list of supplements])
supplements = dict()

count = 0
n_supplements = DB_supplements.shape[0]

# Crawl DB_supplements
for index, row in DB_supplements.iterrows():
    count += 1
    print('Merging metadata for supplemental agreement',count,'out of',n_supplements)

    ID_contract = row['ID_zmluva']

    header_import   = ['Nazov','ID_supplement','Inner-ID','Objednavatel','Dodavatel','Datum_podpisu','Datum_platnosti','Datum_ucinnosti','Poznamka','Prilohy']
    supplement_data = []

    for item in header_import:
        supplement_data.append(row[item])

    supplement_data.append('https://www.crz.gov.sk/index.php?ID='+str(row['ID_supplement'])+'&l=sk')

    if ID_contract in supplements:
        supplements[ID_contract].append(supplement_data)
    else:
        supplements[ID_contract] = [supplement_data]

n_supplements = len(supplements)
not_find = 0

for index, ID_contract in enumerate(supplements):
    print('Saving data for contract ',index+1,'out of',n_supplements)
    location = DB.index[DB['ID'] == ID_contract]
    if len(location) == 0:
        not_find += 1
    else:
        DB.at[location[0], 'Dodatky'] = supplements[ID_contract]

print(not_find,'contracts not merged since contracts are from corrupted XML files ...')
DB.to_csv('CRZ_DB_with_supplements.csv', sep = '|')
